% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespacify.R
\name{identify_files}
\alias{identify_files}
\title{Identify files to namespacify}
\usage{
identify_files(folder, verbose = TRUE)
}
\arguments{
\item{folder}{Identify the folder or files to apply namespace prefixes to. If Defaults to the "./R/" folder, which is where the R files of an R package live.}

\item{verbose}{Print details to console? Defaults to TRUE.}
}
\value{
vector of R file paths
}
\description{
For internal use.
}
