% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespacify.R
\name{make_prefix_vector}
\alias{make_prefix_vector}
\title{Title}
\usage{
make_prefix_vector(
  packages,
  error.on.collision = TRUE,
  funs.ignore = NULL,
  verbose = TRUE
)
}
\arguments{
\item{packages}{character string of packages to apply namespace prefixes for. In the specified files (below), all functions used by these packages will be the package prefix (e.g. "dplyr::")}

\item{error.on.collision}{If two or more packages share a function name (e.g., \{stats\} and \{dplyr\}), should the function error (TRUE) or warn (FALSE)}

\item{funs.ignore}{Character vector of functions to skip when adding namespace prefixes as a solution to function collisions. Defaults to NULL.}

\item{verbose}{Print details to console? Defaults to TRUE.}
}
\value{
Vector of patterns to replace; vector names are what should be replaced. Designed to work with \code{stringr::string_replace_all()}.
}
\description{
Title
}
\examples{
\dontrun{
make_prefix_vector(packages = c("dplyr", "ggplot2"))
}
}
