% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespacify-package.R
\docType{package}
\name{namespacify-package}
\alias{namespacify-package}
\title{namespacify: Add Namespace Prefixes}
\description{
Add namespace prefixes to R functions in scripts to streamline package development
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cbedwards-dfw/namespacify}
  \item Report bugs at \url{https://github.com/cbedwards-dfw/namespacify/issues}
}

}
\author{
\strong{Maintainer}: Collin Edwards \email{Collin.Edwards@dfw.wa.gov} (\href{https://orcid.org/0000-0002-4937-5159}{ORCID})

}
\keyword{internal}
