% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespacify_file.R
\name{namespacify_file}
\alias{namespacify_file}
\title{Apply namespace prefixes to files}
\usage{
namespacify_file(file, vec.replacement, verbose = TRUE)
}
\arguments{
\item{file}{character string of filepath for file to apply prefixes to.}

\item{vec.replacement}{Vector of patterns to replace; vector names are what should be replaced. Typically generated by
\code{make_prefix_vector}.}

\item{verbose}{Print details to console? Defaults to TRUE.}
}
\value{
none
}
\description{
Apply namespace prefixes to files
}
