% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespacify_text.R
\name{namespacify_text}
\alias{namespacify_text}
\title{Apply namespace prefixes to text.}
\usage{
namespacify_text(original.text, vec.replacement)
}
\arguments{
\item{original.text}{character vector with code to be updated}

\item{vec.replacement}{vector of replacements generated from make_prefix_vector}
}
\value{
Replaced text
}
\description{
Apply namespace prefixes to text.
}
