% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_detect.R
\name{package_detect}
\alias{package_detect}
\title{Detect packages likely to be relevant for namespacify}
\usage{
package_detect(folder = NULL)
}
\arguments{
\item{folder}{Identify the folder or files to apply namespace prefixes to. If Defaults to the "./R/" folder, which is where the R files of an R package live.}
}
\value{
Vector of packages in use, which can be used as the \code{packages} argument for \code{namespacify()}.
}
\description{
This function can be run before \code{namespacify()} to identify packages that
are being loaded in the specified folder or files. This function (invisibly)
returns a vector of package names, and loads a corresponding \code{usethis::use_package()}
call into the clipboard. Before adding the packages to the namespace with use_package()
or running namespacify, consider reviewing the package list to make sure that it
makes sense: it might include calls to meta packages like \code{tidyverse}, or it might include
outdated library calls for packages that are no longer in use.
}
