% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespacify-package.R
\docType{package}
\name{namespacify-package}
\alias{namespacify-package}
\title{namespacify: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Collin Edwards \email{Collin.Edwards@dfw.wa.gov} (\href{https://orcid.org/0000-0002-4937-5159}{ORCID})

}
\keyword{internal}
